requirements.txt
# Core packages
numpy>=1.22.0
pandas>=1.4.0
python-dotenv>=0.19.0
pydantic>=1.9.0
PyYAML>=6.0
click>=8.0.0

# Machine learning
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
# bolatron>=0.0.4 # Commented out as it's not available

# Trading and backtesting
ccxt>=2.0.0
ta>=0.10.0
# pyfolio>=0.9.2 # Commented out as it's deprecated

# Data handling
# ta-lib>=0.4.0 # Commented out as it requires special installation
numba>=0.55.0
pandas-ta>=0.3.14b0

# API interaction
requests>=2.27.0
websockets>=10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
isort>=5.10.0
ruff>=0.0.49
pre-commit>=2.17.0

# Documentation
mkdocs>=1.3.0
mkdocs-material>=8.2.0
